+++ components/header.module.css
@@ -1,0 +1,4 @@
+.site_name {
+  font-size: 1rem;
+  letter-spacing: 0.15rem;
+}
